requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow==10.1.0
pygame==2.5.2